extract logic to pluginsConfig
